Use CrossVersion.patch for Silencer plugin
For details on how these differ see:
https://www.scala-sbt.org/1.x/docs/Cross-Build.html#More+about+using+cross-built+libraries

The motivaton for this is so this part of the build doesn't need to be
(ironically) patched when running with a binary-compatible build of the
next patch release of Scala (e.g. Scala 2.13.2-bin-SHA).

Ideally this could be backported to the release-2.5 branch so it would
be immediately of use, otherwise it'll only be available in the
community builds that track Akka 2.6.